Update input.geos for new SOA options and restart filenames
In v9-02o, we have implemented the updated SOA simulation and optional
semivolatile POA updates submitted by Havala Pye. Options to turn on the
semivolatile POA scheme, set napthalene emissions, and scale POA emissions
have been added to the aerosol menu. These options are not used for the
66-tracer full-chemistry simulation.

Lines have also been added to input.geos for users to specify the location
and name of the soil NOx and CSPEC restart files. This will facilitate
work on the GEOS-Chem Unit Tester.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>